Automatisierte Map Generierung und Kamera Funktion. Neues Player Movement.

+++ game/objects/tiles/Air.java
@@ -4,7 +4,7 @@ import java.awt.Graphics2D;
 
 public class Air extends Tile{
 
-	public Air(int x, int y) {
+	public Air(double x, double y) {
 		super(x, y);
 	}
